Adjust: moved local database and secretdevkey to configuration, moved success msg to constants
