+++ utils/configuration.js
@@ -1,0 +1,7 @@
+const local = 'mongodb://localhost:27017/news-local';
+const secretDevKey = 'dev-secret-key';
+
+module.exports = {
+  local,
+  secretDevKey,
+};
