+++ utils/constants.js
@@ -1,6 +1,3 @@
-const local = 'mongodb://localhost:27017/news-local';
-const secretDevKey = 'dev-secret-key';
-
 const SUCCESS_MSG = 201;
 
-module.exports = { local, secretDevKey, SUCCESS_MSG };
+module.exports = { SUCCESS_MSG };
